metrics/requirements.txt
owslib
boto3
requests
numpy
Pillow